Move SpringFox recipes out of SpringDoc package
