+++ src/test/java/org/openrewrite/java/spring/doc/UpgradeSpringDoc2Test.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.java.springdoc;
+package org.openrewrite.java.spring.doc;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
